requirements.txt
openai>=1.40.0
requests>=2.32.3
trafilatura>=1.7.0
rank-bm25>=0.2.2
python-dotenv>=1.0.1
tiktoken>=0.7.0
tenacity>=8.5.0
streamlit>=1.36.0
